requirements.txt
llama-cpp-python==0.3.14
google-api-python-client==2.177.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
python-telegram-bot==22.3
pydantic==2.11.7
python-dateutil>=2.8.0
python-dotenv>=1.0.0
fire==0.7.0
